requirements.txt
awscli==1.11.146
certifi==2018.4.16
chemvae==1.0.0
cycler==0.10.0
docutils==0.14
h5py==2.8.0
jmespath==0.9.3
Keras==2.0.7
kiwisolver==1.0.1
matplotlib==3.0.1
mock==2.0.0
numpy==1.15.3
olefile==0.46
pandas==0.23.4
pbr==4.2.0
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.3.3
pyparsing==2.2.2
python-dateutil==2.6.1
pytz==2018.7
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.11
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tensorflow==1.1.0
Theano==1.0.3
tornado==5.1.1
absl-py==0.1.11
alabaster==0.7.10
anaconda-client==1.7.2
anaconda-navigator==1.8.7
anaconda-project==0.6.0
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astor==0.6.2
astroid==1.4.9
astropy==1.3.2
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.1
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
boto3==1.9.8
botocore==1.12.8
Bottleneck==1.2.1
bz2file==0.98
cffi==1.10.0
chardet==3.0.3
click==6.7
click-plugins==1.0.4
cligj==0.4.0
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
conda==4.5.11
conda-build==3.16.2
contextlib2==0.5.5
cryptography==1.8.1
cymem==1.31.2
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.4
decorator==4.3.0
descartes==1.1.0
distributed==1.16.3
entrypoints==0.2.2
et-xmlfile==1.0.1
fastai==0.7.0
fastcache==1.0.2
feather-format==0.4.0
filelock==3.0.10
Fiona==1.7.13
Flask==0.12.2
Flask-Cors==3.0.2
gast==0.2.0
gensim==3.6.0
geopandas==0.4.0
gevent==1.2.1
glob2==0.6
graphviz==0.9
greenlet==0.4.12
grpcio==1.10.0
gunicorn==19.8.1
HeapDict==1.0.0
html5lib==0.9999999
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
isoweek==1.3.3
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras-Applications==1.0.4
keras-contrib==2.0.8
Keras-Preprocessing==1.0.2
keras-tqdm==2.0.1
lazy-object-proxy==1.2.2
libarchive-c==2.8
llvmlite==0.18.0
lmdb==0.94
locket==0.2.0
lxml==3.7.3
Markdown==2.6.11
MarkupSafe==0.23
mistune==0.7.4
mizani==0.4.6
mkl-fft==1.0.6
mkl-random==1.0.2
mordred==1.1.2
mpmath==0.19
msgpack==0.5.6
msgpack-numpy==0.4.3.2
msgpack-python==0.4.8
multipledispatch==0.4.9
munch==2.3.2
murmurhash==0.28.0
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==2.1
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.5
numpydoc==0.6.0
odo==0.5.0
opencv-python==3.4.0.12
openpyxl==2.4.7
packaging==16.8
palettable==3.1.1
pandas-summary==0.0.5
pandocfilters==1.4.1
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
pipenv==11.10.0
pkginfo==1.4.2
plotly==3.1.1
plotnine==0.4.0
ply==3.10
preshed==1.0.1
prompt-toolkit==1.0.14
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.33
pyarrow==0.10.0
pycairo==1.18.0
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyproj==1.9.5.1
pytest==3.0.7
PyWavelets==0.5.2
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.5.0
requests==2.14.2
retrying==1.3.3
rope-py3k==0.9.4.post1
scikit-image==0.13.0
seaborn==0.7.1
selenium==3.5.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
singledispatch==3.4.0.3
sklearn-pandas==1.7.0
smart-open==1.7.1
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
sympy==1.0
tables==3.3.0
tblib==1.3.2
tensorboard==1.6.0
termcolor==1.1.0
terminado==0.6
testpath==0.3
toolz==0.8.2
torch==0.3.1
torchtext==0.2.3
torchvision==0.2.1
tqdm==4.28.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.23
virtualenv==15.2.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==2.0.0
wrapt==1.10.10
xgboost==0.80
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
